+++ src/CfgParserVarExpander.hh
@@ -1,0 +1,33 @@
+//
+// CfgParserVarExpander.hh for pekwm
+// Copyright (C) 2022 Claes Nästén <pekdon@gmail.com>
+//
+// This program is licensed under the GNU GPL.
+// See the LICENSE file for more information.
+//
+
+#ifndef _PEKWM_CFG_PARSER_VAR_EXPANDER_HH_
+#define _PEKWM_CFG_PARSER_VAR_EXPANDER_HH_
+
+#include <string>
+
+enum CfgParserVarExpanderType {
+	CFG_PARSER_VAR_EXPANDER_MEM,
+	CFG_PARSER_VAR_EXPANDER_OS_ENV,
+	CFG_PARSER_VAR_EXPANDER_X11_ATOM,
+	CFG_PARSER_VAR_EXPANDER_X11_RES
+};
+
+class CfgParserVarExpander {
+public:
+	virtual ~CfgParserVarExpander() { }
+
+	virtual void clear() { }
+	virtual void define(const std::string& name,
+			    const std::string& val) { }
+	virtual bool lookup(const std::string& name, std::string& val) = 0;
+};
+
+CfgParserVarExpander* mkCfgParserVarExpander(CfgParserVarExpanderType type);
+
+#endif // _PEKWM_CFG_PARSER_VAR_EXPANDER_HH_